Added animations to NPC's and shooting trigger, with base health decreasing
Added animations to the enemy NPC.

Added a trigger so the animation for the NPC will stop and change to a different image for the NPC and the trigger will also make the NPC start shooting at the base.

The health for the base has been set to decrease wehn the NPC's hit the base.